MtgPula.Tournaments (mtg_pula v0.1.0)

The Tournaments context.

Functions

calculate_gw(player)

Calculates average opponents game win procentage from each game win that appears on certain tournament

calculate_ogp(player)

Calculates average opponents game win procentage from players list of opponents

calculate_points(player)

Calculates points for a player

Examples

iex> calculate_points(player)
%Player{}

calculate_procentage_omw(player, tournament)

Calculates average opponents match win procentage from players list of opponents

Examples

calculate_tiebreakers(standings, tournament)

Calculates tiebreakers for a certain standings and tournament

Examples

iex> calculate_tiebreakers(standings, tournament)
[%Player{}, ...]

change_match(match, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking match changes.

Examples

iex> change_match(match)
%Ecto.Changeset{data: %Match{}}

change_player(player, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking player changes.

Examples

iex> change_player(player)
%Ecto.Changeset{data: %Player{}}

change_tournament(tournament, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking tournament changes.

Examples

iex> change_tournament(tournament)
%Ecto.Changeset{data: %Tournament{}}

create_match(attrs \\ %{})

create_player(attrs \\ %{})

Creates a player.

Examples

iex> create_player(%{field: value})
{:ok, %Player{}}

iex> create_player(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_tournament(attrs \\ %{})

Creates a tournament.

Examples

iex> create_tournament(%{field: value})
{:ok, %Tournament{}}

iex> create_tournament(%{field: bad_value})
{:error, %Ecto.Changeset{}}

current_matches(tournament_id)

Finds and returns the current round matches

delete_match(match)

Deletes a match.

Examples

iex> delete_match(match)
{:ok, %Match{}}

iex> delete_match(match)
{:error, %Ecto.Changeset{}}

delete_player(player_id)

Deletes a player.

Examples

iex> delete_player(player)
{:ok, %Player{}}

iex> delete_player(player)
{:error, %Ecto.Changeset{}}

delete_tournament(tournament)

Deletes a tournament.

Examples

iex> delete_tournament(tournament)
{:ok, %Tournament{}}

iex> delete_tournament(tournament)
{:error, %Ecto.Changeset{}}

drop_player(player_id)

find_pair(player, rest)

Goes through all the players that
are left to match them to a player that hasnt been paired with already.

get_match!(id)

Gets a single match.

Raises Ecto.NoResultsError if the Match does not exist.

Examples

iex> get_match!(123)
%Match{}

iex> get_match!(456)
** (Ecto.NoResultsError)

get_player!(id)

Gets a single player.

Raises Ecto.NoResultsError if the Player does not exist.

Examples

iex> get_player!(123)
%Player{}

iex> get_player!(456)
** (Ecto.NoResultsError)

get_tournament!(id)

Gets a single tournament.

Raises Ecto.NoResultsError if the Tournament does not exist.

Examples

iex> get_tournament!(123)
%Tournament{}

iex> get_tournament!(456)
** (Ecto.NoResultsError)

get_tournament_by_join_code(join_code)

Gets a tournament by join code.

Examples

iex> get_tournament_by_join_code("123456")
%Tournament{}

iex> get_tournament_by_join_code("123456")
nil

list_all_tournament_players(tournament_id)

Lists all players added to a certain tournament

Examples

iex> list_all_tournament_players(123)
[%Player{}, ...]

list_matches()

Returns the list of matches.

Examples

iex> list_matches()
[%Match{}, ...]

list_players()

Returns the list of players.

Examples

iex> list_players()
[%Player{}, ...]

list_tournaments()

Returns the list of tournaments.

Examples

iex> list_tournaments()
[%Tournament{}, ...]

list_tournaments_with_players_and_users()

Lists tournaments with their players

Examples

make_pairings(list, pairings)

Recursively uses find_pair, to pair players by their performance. If there is a uneven number of players it gives the player that already hasnt had a bye, a bye, if all players had a bye it takes the first player that cant be paired. It returns an array of tuples of match pairs.

pair_next_round(tournament_id)

Takes in tournament_id. Calculates current standings based on player performance. Updates the current_round, to match the round we are currently in. Returns pairings for the current round.

prepare_matches(tournament_id)

Takes in tournament_id. Pairs next round, through pair_next_round() (see pair_next_round/1) and creates matches in the database

standings_on_tournament(tournament_id)

Returns standings with tiebreakers and points

Examples

iex> standings_on_tournament(123)
[%Player{}, ...]

update_match(match, attrs)

Updates a match.

Examples

iex> update_match(match, %{field: new_value})
{:ok, %Match{}}

iex> update_match(match, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_played(attrs)

Creates a match.

Examples

iex> create_match(%{field: value})
{:ok, %Match{}}

iex> create_match(%{field: bad_value})
{:error, %Ecto.Changeset{}}

update_player(player, attrs)

Updates a player.

Examples

iex> update_player(player, %{field: new_value})
{:ok, %Player{}}

iex> update_player(player, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_tournament(tournament, attrs)

Updates a tournament.

Examples

iex> update_tournament(tournament, %{field: new_value})
{:ok, %Tournament{}}

iex> update_tournament(tournament, %{field: bad_value})
{:error, %Ecto.Changeset{}}