MtgPula.Tournaments.Player (mtg_pula v0.1.0)

Represents a player schema.

This schema is used to store player-related information, such as points, deck, and match history.

Fields

id (binary_id): The unique identifier for the player (autogenerated).

deck (string): The deck used by the player.

opponents (array of UUIDs): The list of opponents the player has faced.

points (integer): The total points earned by the player.

had_bye (boolean): Indicates if the player had a bye in any round.

dropped (boolean): Indicates if the player has dropped out of the tournament.

omw (float, virtual): Opponent match win percentage.

gw (float, virtual): Game win percentage.

ogp (float, virtual): Opponent game percentage.

inserted_at (utc_datetime): The timestamp when the player was created.

updated_at (utc_datetime): The timestamp when the player was last updated.

Relationships

belongs_to :tournament: The tournament the player is participating in. See Tournament Schema.

belongs_to :user: The user associated with the player. See User Schema.

has_many :matches_as_player1: Matches where the player is player 1. See Match Schema.

has_many :matches_as_player2: Matches where the player is player 2. See Match Schema.

has_many :matches_as_winner: Matches where the player is the winner. See Match Schema.

Examples

# Creating a new player
attrs = %{deck: "Mono Red Aggro", tournament_id: "tournament_uuid", user_id: "user_uuid"}
changeset = MtgPula.Tournaments.Player.changeset(%MtgPula.Tournaments.Player{}, attrs)

if changeset.valid? do
  Repo.insert(changeset)
else
  IO.inspect(changeset.errors)
end

Functions

changeset(player, attrs)

Creates a changeset for a player.

Parameters

player (struct): The player struct to update.

attrs (map): A map of attributes to apply to the changeset.

Validations

Ensures required fields are present.

Enforces unique constraints for user and tournament associations.

Examples

iex> changeset(%MtgPula.Tournaments.Player{}, %{deck: "Mono Red Aggro"})
%Ecto.Changeset{...}

iex> changeset(%MtgPula.Tournaments.Player{}, %{deck: nil})
%Ecto.Changeset{valid?: false, errors: [deck: {"can't be blank", []}]}