MtgPula.Tournaments.Tournament (mtg_pula v0.1.0)

Represents a tournament schema.

This schema is used to store tournament-related information, such as name, rounds, and players.

Fields

id (binary_id): The unique identifier for the tournament (autogenerated).

name (string): The name of the tournament.

finished (boolean): Indicates if the tournament is finished.

current_round (integer): The current round of the tournament.

number_of_rounds (integer): The total number of rounds in the tournament.

join_code (string): A unique code for players to join the tournament.

inserted_at (utc_datetime): The timestamp when the tournament was created.

updated_at (utc_datetime): The timestamp when the tournament was last updated.

Relationships

has_many :match: The matches associated with the tournament. See Match Schema.

has_many :player: The players participating in the tournament. See Player Schema.

belongs_to :user: The user who created the tournament. See User Schema.

Examples

# Creating a new tournament
attrs = %{name: "Summer Cup", number_of_rounds: 5}
changeset = MtgPula.Tournaments.Tournament.changeset(%MtgPula.Tournaments.Tournament{}, attrs)

if changeset.valid? do
  Repo.insert(changeset)
else
  IO.inspect(changeset.errors)
end

Functions

changeset(tournament, attrs)

Creates a changeset for a tournament.

Parameters

tournament (struct): The tournament struct to update.

attrs (map): A map of attributes to apply to the changeset.

Validations

Ensures required fields are present.

Generates a unique join code if not provided.

Examples

iex> changeset(%MtgPula.Tournaments.Tournament{}, %{name: "Summer Cup"})
%Ecto.Changeset{...}

iex> changeset(%MtgPula.Tournaments.Tournament{}, %{name: nil})
%Ecto.Changeset{valid?: false, errors: [name: {"can't be blank", []}]}